해쉬태그,포스트태그 기능 1차 완성

+++ src/main/java/com/example/blogex/domain/follow/repository/FollowRepository.java
@@ -19,10 +19,10 @@ public interface FollowRepository extends JpaRepository<Follow, Long> {
     boolean exitsByFollowerIdAndFollowingUserId(Long followerId, Long followingUserId);
 
     //내가 얼마나 팔로우하고 있는지 카운트(내가 follower)
-    long countByFollowerId(Long id);
+    int countByFollowerId(Long id);
 
     //나를 얼마나 팔로잉하고 있는지 카운트
-    long countByFollowingUserId(Long id);
+    int countByFollowingUserId(Long id);
 
     void deleteByFollowerIdAndFollowingUserId(Long followerId, Long followingUserId);
